requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
chardet==3.0.4
colour-demosaicing==0.1.4
colour-science==0.3.13
configparser==4.0.2
cycler==0.10.0
futures==3.3.0
idna==2.8
jsonpatch==1.24
jsonpointer==2.0
kiwisolver==1.1.0
matplotlib==2.2.4
numpy==1.16.5
opencv-python==4.1.2.30
pillow==6.1.0
protobuf==3.11.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
requests==2.22.0
scipy==1.2.2
singledispatch==3.4.0.3
six==1.12.0
subprocess32==3.5.4
tensorboardx==1.6
torchfile==0.1.0
torchvision==0.2.0
tornado==5.1.1
transforms3d==0.3.1
urllib3==1.25.4
visdom==0.1.8.9
websocket-client==0.56.0
